requirements.txt
pandas
requests
google-cloud-bigquery
google-auth